Move loot crate group logic into loot_crates.py
